fix: Cambié los tipos de datos Bigint a Long + añadí autoincremento de IDs + añadí constructores

+++ demo/src/main/java/com/example/demo/AsignaturaHabilidadRepository.java
@@ -2,7 +2,6 @@ package com.example.demo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigDecimal;
 
-public interface AsignaturaHabilidadRepository extends JpaRepository<AsignaturaHabilidad, BigDecimal> {
+public interface AsignaturaHabilidadRepository extends JpaRepository<AsignaturaHabilidad, Long> {
 }
