+++ demo/src/main/java/com/example/demo/AulaRepository.java
@@ -2,7 +2,5 @@ package com.example.demo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigDecimal;
-
-public interface AulaRepository extends JpaRepository<Aula, BigDecimal> {
+public interface AulaRepository extends JpaRepository<Aula, Long> {
 }
