+++ demo/src/main/java/com/example/demo/ColegioRepository.java
@@ -2,7 +2,5 @@ package com.example.demo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigDecimal;
-
-public interface ColegioRepository extends JpaRepository<Colegio, BigDecimal> {
+public interface ColegioRepository extends JpaRepository<Colegio, Long> {
 }
