+++ demo/src/main/java/com/example/demo/EditorialRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigDecimal;
 
-public interface EditorialRepository extends JpaRepository<Editorial, BigDecimal> {
+public interface EditorialRepository extends JpaRepository<Editorial, Long> {
 }
