+++ demo/src/main/java/com/example/demo/LibroRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigDecimal;
 
-public interface LibroRepository extends JpaRepository<Libro, BigDecimal> {
+public interface LibroRepository extends JpaRepository<Libro, Long> {
 }
