+++ demo/src/main/java/com/example/demo/ProfesorRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigDecimal;
 
-public interface ProfesorRepository extends JpaRepository<Profesor, BigDecimal> {
+public interface ProfesorRepository extends JpaRepository<Profesor, Long> {
 }
